# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Emerald
  description: Memory and context infrastructure for AI agents
  version: 0.3.0
  contact:
    name: Emerald
    url: https://emerald.ai
  license:
    name: MIT
paths:
...
  /v1/upload:
    post:
      tags:
      - Upload
      summary: Upload File
      description: 'Upload a file (PDF, image, audio, video, code, etc.) for async

        extraction, chunking, embedding, and indexing. Returns 202 Accepted

        with a pipeline_id to poll via ``GET /v1/pipelines/{id}``.


        File size limit: 50MB.'
      operationId: upload_file_v1_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_v1_upload_post'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Upload File V1 Upload Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/files:
    get:
      tags:
      - Upload
      summary: List Files
      description: List uploaded files for an entity with cursor-based pagination.
      operationId: list_files_v1_files_get
      parameters:
      - name: entity_id
        in: query
        required: true
        schema:
          type: string
          title: Entity Id
      - name: status_filter
        in: query
        required: false
        schema:
          type: string
          default: done
          title: Status Filter
      - name: page_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Cursor token for pagination
          title: Page Token
        description: Cursor token for pagination
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Page Size
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Files V1 Files Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v2/upload:
    post:
      tags:
      - Upload
      summary: Upload File
      description: 'Upload a file (PDF, image, audio, video, code, etc.) for async

        extraction, chunking, embedding, and indexing. Returns 202 Accepted

        with a pipeline_id to poll via ``GET /v1/pipelines/{id}``.


        File size limit: 50MB.'
      operationId: upload_file_v2_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_v2_upload_post'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Upload File V2 Upload Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v2/files:
    get:
      tags:
      - Upload
      summary: List Files
      description: List uploaded files for an entity with cursor-based pagination.
      operationId: list_files_v2_files_get
      parameters:
      - name: entity_id
        in: query
        required: true
        schema:
          type: string
          title: Entity Id
      - name: status_filter
        in: query
        required: false
        schema:
          type: string
          default: done
          title: Status Filter
      - name: page_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Cursor token for pagination
          title: Page Token
        description: Cursor token for pagination
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Page Size
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Files V2 Files Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_file_v1_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        entity_id:
          type: string
          title: Entity Id
        content_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Type
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - file
      - entity_id
      title: Body_upload_file_v1_upload_post
    Body_upload_file_v2_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        entity_id:
          type: string
          title: Entity Id
        content_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Type
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - file
      - entity_id
      title: Body_upload_file_v2_upload_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: em_<32-hex>
      description: All endpoints require a Bearer API key in the `Authorization` header. Keys are prefixed
        with `em_` and scoped to a single entity. The server stores only the SHA-256 hash.
servers:
- url: http://localhost:8000/v1
  description: Local development
- url: https://api.emerald.ai/v1
  description: Production
...
security:
- ApiKeyAuth: []